VideoFromJSONAPI/requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
ffmpeg==1.4
ffprobe==0.5
Flask>=2.0.1
Flask-Cors==5.0.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
matplotlib>=3.7.0
moviepy>=1.0.3
numpy>=1.21.0
pillow>=8.3.1
proglog==0.1.10
python-dotenv>=0.19.0
requests>=2.26.0
setuptools==75.6.0
tqdm==4.67.0
urllib3==2.2.3
Werkzeug==3.1.3
pytest>=6.2.5
librosa==0.11.0
scipy>=1.6.0
pooch>=1.1
typing_extensions>=4.1.1
audioread>=2.1.9
scikit-learn>=1.1.0
lazy_loader>=0.1
soundfile>=0.12.1
supabase>=2.0.0
